refactor: serverHandler initialization
